Added function playerSyncAllData in JUPPController for optimize calls to JUPP. Some optimizations. Changed JUPP version to 0.0.2. Changed all mentions of 'int playerID' to 'short playerID'. Started commits at UPS and Client. Todo: Add smooth sync players to client; Add window with information about connect to server and button reconnect.

+++ src/main/java/com/VER7U7/Server/Packets/Handlers/NewIncomingConnectionPacket.java
@@ -12,7 +12,7 @@ public class NewIncomingConnectionPacket implements PacketFunction {
     private FunctionGlobalArgs globalArgs;
 
 
-    public void process(int playerID, NetworkPacket networkPacket) {
+    public void process(short playerID, NetworkPacket networkPacket) {
 
 
     }
