+++ src/main/java/com/VER7U7/Server/Packets/Handlers/PacketFunction.java
@@ -5,7 +5,7 @@ import com.VER7U7.Server.Network.NetworkPacket;
 import static com.VER7U7.Server.Packets.Factory.PacketConstants.*;
 
 public interface PacketFunction {
-    public void process(int playerID, NetworkPacket networkPacket);
+    public void process(short playerID, NetworkPacket networkPacket);
 
     //need return id of packet
     public IncomingPacketType initialize(FunctionGlobalArgs globalArgs);
